2.0.0  download download 第二版 1.每隔一段时间根据读取时数据量计算网速 2.在写入数据时阻塞 实现下载速度控制

+++ okdownload/src/main/java/com/luckyaf/okdownload/speedContral/SpeedCallback.java
@@ -7,5 +7,5 @@ package com.luckyaf.okdownload.speedContral;
  */
 public interface SpeedCallback {
 
-    void call(float speed);
+    void call(long speed);
 }
